+++ src/activities/reader/MdReaderMenuActivity.cpp
@@ -36,12 +36,12 @@ void MdReaderMenuActivity::onExit() { Activity::onExit(); }
 
 void MdReaderMenuActivity::loop() {
   buttonNavigator.onNext([this] {
-    selectedIndex = ButtonNavigator::nextIndex(selectedIndex, static_cast<int>(menuItems.size()));
+    selectedIndex = ButtonNavigator::nextIndex(selectedIndex, menuItems.size());
     requestUpdate();
   });
 
   buttonNavigator.onPrevious([this] {
-    selectedIndex = ButtonNavigator::previousIndex(selectedIndex, static_cast<int>(menuItems.size()));
+    selectedIndex = ButtonNavigator::previousIndex(selectedIndex, menuItems.size());
     requestUpdate();
   });
